+++ msg4j-core/src/main/java/com/dopsun/msg4j/core/WritableMessageImpl.java
@@ -98,7 +98,7 @@ final class WritableMessageImpl extends AbstractMessage implements WritableMessa
     }
 
     @Override
-    public void putLong(String fieldName, int value) throws InvalidTypeException {
+    public void putLong(String fieldName, long value) throws InvalidTypeException {
         Objects.requireNonNull(fieldName);
 
         Field field = fields().get(fieldName);
